+++ tasks/tests/tasks_tests/test_teardown/neverending/tasks.py
@@ -1,0 +1,15 @@
+import time
+
+from robocorp.tasks import task, task_cache
+
+
+@task_cache
+def some_resource():
+    yield "resource"
+    while True:
+        time.sleep(1)
+
+
+@task
+def neverending():
+    some_resource()